requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
langchain
langgraph
chromadb
sentence-transformers
python-dotenv
groq
pypdf